model/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.8.1
numpy==1.26.1
python-dotenv==1.0.0
scipy==1.11.4
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
